4860 data dog sketch (#4866)
## What is the purpose of the change

This PR adds a DD-Sketch and a new metric, quantile.

## Verifying this change

This change adds tests.

## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): no
  - The compiler: no
  - The threading model: no
  - The Runtime per-record code paths (performance sensitive): yes
  - The network stack: no
- Anything that affects deployment or recovery: Coordinator (and its
components), NodeEngine (and its components): no

## Documentation

  - Does this pull request introduce a new feature? no

## Issue Closed by this pull request:

This PR closes #4860

---------

+++ nes-execution/tests/UnitTests/Execution/Expressions/Functions/AbsExpressionTest.cpp
@@ -40,7 +40,7 @@ TEST_F(AbsExpressionTest, evaluateAbsExpressionFloat) {
     {
         auto resultValue = expression.eval(Value<Float>((float) -17.4));
         ASSERT_EQ(resultValue, (float) 17.4);
-        ASSERT_TRUE(resultValue->getTypeIdentifier()->isType<Double>());
+        ASSERT_TRUE(resultValue->getTypeIdentifier()->isType<Float>());
     }
 
     // Double
